Reorganize css files in pages/home
