+++ src/pages/home/index.css
@@ -1,129 +1,4 @@
-.video-background {
-    position: relative;
-    width: 100%;
-    height: 70vh;
-    overflow: hidden;
-}
-
-.video {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    min-width: 100%;
-    min-height: 100%;
-    width: auto;
-    height: auto;
-    z-index: -1000;
-    transform: translate(-50%, -50%);
-    background-size: cover;
-    transition: 1s opacity;
-}
-
-.wd-content {
-    position: relative;
-    z-index: 10;
-}
-
-.wd-vline {
-    border-left: 1px black solid;
-    height: 300px;
-}
-
 .wd-text-container {
     display: flex;
     align-items: flex-start;
-}
-
-.wd-relative {
-    position: relative;
-}
-
-
-.slider {
-    margin: 0 20px;
-    overflow: "hidden";
-    padding: 2rem 0;
-}
-
-.slider img {
-    width: 100%;
-    border-radius: 10px;
-}
-
-.react-multi-carousel-list {
-    padding: 0rem 0 2rem 0;
-}
-
-.custom-dot-list-style button {
-    border: none;
-    background: rgb(255, 68, 68);
-}
-
-.react-multi-carousel-dot.react-multi-carousel-dot--active button {
-    background: rgb(255, 68, 68) !important;
-}
-
-.parent {
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
-}
-
-.wd-purple-background {
-    background-color: rgb(204, 173, 234);
-}
-
-.wd-white-background {
-    background-color: white;
-}
-
-/* 
-.outer-div {
-    background-color: rgb(204, 173, 234);
-    padding: 50px;
-}
-
-.inner-div {
-    background-color: white;
-    padding: 50px;
-}
-
-.block {
-    display: inline-block;
-    width: 50%; 
-} */
-
-.outer-div {
-    background-color: rgb(204, 173, 234);
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: 100vh; 
-    padding: 20px; 
-}
-
-.inner-div {
-    background-color: white;
-    padding: 50px;
-    text-align: center;
-    width: 60%; 
-    margin-top: 25px;
-    margin-bottom: 25px; 
-}
-
-.title {
-    margin-bottom: 20px;
-}
-
-.search-bar {
-    display: flex;
-    justify-content: center;
-}
-
-.search-bar input[type="text"] {
-    flex-grow: 1;
-    padding: 10px;
-    border-radius: 5px;
-    border: 1px solid #ccc;
 }
