+++ src/pages/home/video/index.css
@@ -1,0 +1,25 @@
+.wd-video-background {
+    position: relative;
+    width: 100%;
+    height: 70vh;
+    overflow: hidden;
+}
+
+.wd-video {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    min-width: 100%;
+    min-height: 100%;
+    width: auto;
+    height: auto;
+    z-index: -1000;
+    transform: translate(-50%, -50%);
+    background-size: cover;
+    transition: 1s opacity;
+}
+
+.wd-vline {
+    border-left: 1px black solid;
+    height: 300px;
+}
